frontend/api/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyJWT>=2.10.0
bcrypt>=4.0.0
